Move client/lcd to lite
