+++ lite/statik/init.go
