+++ lite/version.go
@@ -1,8 +1,9 @@
-package lcd
+package lite
 
 import (
 	"fmt"
 	"net/http"
+
 	"github.com/irisnet/irishub/client/context"
 	"github.com/irisnet/irishub/version"
 )
@@ -16,7 +17,7 @@ func CLIVersionRequestHandler(w http.ResponseWriter, r *http.Request) {
 // connected node version REST handler endpoint
 func NodeVersionRequestHandler(cliCtx context.CLIContext) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		version, err := cliCtx.Query("/app/version",nil)
+		version, err := cliCtx.Query("/app/version", nil)
 		if err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
 			w.Write([]byte(fmt.Sprintf("Could't query version. Error: %s", err.Error())))
